fix(build): make GCC-specific warnings conditional and fix narrowing conversion
- Make -Wduplicated-cond, -Wduplicated-branches, and -Wlogical-op
  GCC-only since Clang doesn't support these flags
- Fix narrowing conversion error in test_discrete_field.cpp by
  explicitly casting int to size_t in initializer list

+++ tests/unit/core/test_discrete_field.cpp
@@ -116,7 +116,7 @@ TEST_CASE("pfc::interpolate() free function works correctly",
 
     // Initialize: field[i] = i
     for (int i = 0; i < 5; i++) {
-      field[{i}] = i;
+      field[{static_cast<size_t>(i)}] = i;
     }
 
     // Test rounding to nearest (std::round behavior)
